+++ WpfApp5/WpfApp5/Form3.cs
@@ -28,20 +28,27 @@ namespace WpfApp5
 
         private void button1_Click(object sender, EventArgs e)
         {
-            con.Open();
-            SqlCommand cmd = con.CreateCommand();
-            cmd.CommandType = CommandType.Text;
-            cmd.CommandText = "insert into Pupils values('" + textBox1.Text + "','" + textBox4.Text + "','" + textBox3.Text + "','" + textBox2.Text + "','" + textBox5.Text + "')";
-            cmd.ExecuteNonQuery();
-            con.Close();
-            disp_data();
-            textBox1.Clear();
-            textBox2.Clear();
-            textBox3.Clear();
-            textBox4.Clear();
-            textBox5.Clear();
+            DialogResult dialog = MessageBox.Show("Вы уверены, что хотите сохранить?","Сохранить", MessageBoxButtons.YesNo);
+            if (dialog == DialogResult.Yes)
+            {
+                con.Open();
+                SqlCommand cmd = con.CreateCommand();
+                cmd.CommandType = CommandType.Text;
+                cmd.CommandText = "insert into Pupils values('" + textBox1.Text + "','" + textBox4.Text + "','" + textBox3.Text + "','" + textBox2.Text + "','" + textBox5.Text + "')";
+                cmd.ExecuteNonQuery();
+                con.Close();
+                disp_data();
+                textBox1.Clear();
+                textBox2.Clear();
+                textBox3.Clear();
+                textBox4.Clear();
+                textBox5.Clear();
+            }
 
-            MessageBox.Show("Данные внесены");
+            else if(dialog == DialogResult.No)
+            {
+
+            }
         }
 
         private void disp_data()
@@ -101,24 +108,30 @@ namespace WpfApp5
         }
 
         private void button2_Click(object sender, EventArgs e)
-        {
-            con.Open();
-            SqlCommand cmd = con.CreateCommand();
-            cmd.CommandType = CommandType.Text;
-            cmd.CommandText = "delete from Pupils where Name='"+textBox1.Text+"'";
-            cmd.ExecuteNonQuery();
-            con.Close();
-            disp_data();
-            textBox1.Clear();
+        { DialogResult dialog = MessageBox.Show("Вы уверены, что хотите удалить?", "Удалить", MessageBoxButtons.YesNo);
+            if (dialog == DialogResult.Yes)
+            {
+                con.Open();
+                SqlCommand cmd = con.CreateCommand();
+                cmd.CommandType = CommandType.Text;
+                cmd.CommandText = "delete from Pupils where Name='" + textBox1.Text + "'";
+                cmd.ExecuteNonQuery();
+                con.Close();
+                disp_data();
+                textBox1.Clear();
+                MessageBox.Show("Данные удалены");
+            }
 
-            MessageBox.Show("Данные удалены");
+            else if (dialog == DialogResult.No)
+            {
+
+            }
         }
 
         private void button4_Click(object sender, EventArgs e)
         {
             disp_data();
 
-            MessageBox.Show("Данные Выведены");
         }
 
         private void button3_Click(object sender, EventArgs e)
@@ -126,7 +139,7 @@ namespace WpfApp5
             con.Open();
             SqlCommand cmd = con.CreateCommand();
             cmd.CommandType = CommandType.Text;
-            cmd.CommandText = "select * from Pupils where Name='" + textBox1.Text + "'";
+            cmd.CommandText = "select * from Pupils where Name='%" + textBox1.Text + "'";
             cmd.ExecuteNonQuery();
             DataTable dt = new DataTable();
             SqlDataAdapter sda = new SqlDataAdapter(cmd);
