+++ WpfApp5/WpfApp5/Subjects.cs
@@ -55,31 +55,36 @@ namespace WpfApp5
 
 
         private void button1_Click_1(object sender, EventArgs e)
-        {
-            con.Open();
-            SqlCommand cmd = con.CreateCommand();
-            cmd.CommandType = CommandType.Text;
-            cmd.CommandText = "insert into Subjects values('" + textBox1.Text + "')";
-            cmd.ExecuteNonQuery();
-            con.Close();
-            disp_data();
-            textBox1.Clear();
-            
+        { DialogResult dialog = MessageBox.Show("Вы уверены, что хотите сохранить?", "Сохранить", MessageBoxButtons.YesNo);
+            if (dialog == DialogResult.Yes)
+            {
+                con.Open();
+                SqlCommand cmd = con.CreateCommand();
+                cmd.CommandType = CommandType.Text;
+                cmd.CommandText = "insert into Subjects values('" + textBox1.Text + "')";
+                cmd.ExecuteNonQuery();
+                con.Close();
+                disp_data();
+                textBox1.Clear();
 
-            MessageBox.Show("Данные внесены");
+
+            }
         }
 
         private void button2_Click(object sender, EventArgs e)
-        {
-            con.Open();
-            SqlCommand cmd = con.CreateCommand();
-            cmd.CommandType = CommandType.Text;
-            cmd.CommandText = "delete from Subjects where Name='" + textBox1.Text + "'";
-            cmd.ExecuteNonQuery();
-            con.Close();
-            disp_data();
-            textBox1.Clear();
-            MessageBox.Show("Данные удалены");
+        { DialogResult dialog = MessageBox.Show("Вы уверены, что хотите удалить?", "Удалить", MessageBoxButtons.YesNo);
+            if (dialog == DialogResult.Yes)
+            {
+                con.Open();
+                SqlCommand cmd = con.CreateCommand();
+                cmd.CommandType = CommandType.Text;
+                cmd.CommandText = "delete from Subjects where Name='" + textBox1.Text + "'";
+                cmd.ExecuteNonQuery();
+                con.Close();
+                disp_data();
+                textBox1.Clear();
+                MessageBox.Show("Данные удалены");
+            }
         }
 
         private void button4_Click(object sender, EventArgs e)
