+++ WpfApp5/WpfApp5/Teachers.cs
@@ -37,10 +37,7 @@ namespace WpfApp5
 
         }
 
-        private void button1_Click_1(object sender, EventArgs e)
-        {
-            this.teachersTableAdapter.Update(this.schoolDataSet);
-        }
+       
 
         private void disp_data()
         {
@@ -57,34 +54,39 @@ namespace WpfApp5
         }
 
         private void button4_Click(object sender, EventArgs e)
-        {
-            con.Open();
-            SqlCommand cmd = con.CreateCommand();
-            cmd.CommandType = CommandType.Text;
-            cmd.CommandText = "insert into Teachers values('" + textBox1.Text + "','" + textBox4.Text + "','" + textBox3.Text + "','" + textBox2.Text + "','" + textBox5.Text + "','" + textBox6.Text + "')";
-            cmd.ExecuteNonQuery();
-            con.Close();
-            disp_data();
-            textBox1.Clear();
-            textBox2.Clear();
-            textBox3.Clear();
-            textBox4.Clear();
-            textBox5.Clear();
-            textBox6.Clear();
+        { DialogResult dialog = MessageBox.Show("Вы уверены, что хотите сохранить?", "Сохранить", MessageBoxButtons.YesNo);
+            if (dialog == DialogResult.Yes)
+            {
+                con.Open();
+                SqlCommand cmd = con.CreateCommand();
+                cmd.CommandType = CommandType.Text;
+                cmd.CommandText = "insert into Teachers values('" + textBox1.Text + "','" + textBox4.Text + "','" + textBox3.Text + "','" + textBox2.Text + "','" + textBox5.Text + "','" + textBox6.Text + "')";
+                cmd.ExecuteNonQuery();
+                con.Close();
+                disp_data();
+                textBox1.Clear();
+                textBox2.Clear();
+                textBox3.Clear();
+                textBox4.Clear();
+                textBox5.Clear();
+                textBox6.Clear();
+            }
         }
 
         private void button3_Click(object sender, EventArgs e)
-        {
-            con.Open();
-            SqlCommand cmd = con.CreateCommand();
-            cmd.CommandType = CommandType.Text;
-            cmd.CommandText = "delete from Teachers where Name='" + textBox1.Text + "'";
-            cmd.ExecuteNonQuery();
-            con.Close();
-            disp_data();
-            textBox1.Clear();
+        { DialogResult dialog = MessageBox.Show("Вы уверены, что хотите удалить?", "Удалить", MessageBoxButtons.YesNo);
+            if (dialog == DialogResult.Yes)
+            {
+                con.Open();
+                SqlCommand cmd = con.CreateCommand();
+                cmd.CommandType = CommandType.Text;
+                cmd.CommandText = "delete from Teachers where Name='" + textBox1.Text + "'";
+                cmd.ExecuteNonQuery();
+                con.Close();
+                disp_data();
+                textBox1.Clear();
 
-            MessageBox.Show("Данные удалены");
+            }
         }
 
         private void button1_Click(object sender, EventArgs e)
